Set IMA policy for container
Implement support for providing an IMA policy. The policy can either
be provided via the container's Config or if none is provided
the policy is (tried to be) read from the container's filesystem from
their usual distro-specific locations.

The activation of the policy happens between the clone() and the execve()
calls so that the first executable running inside the container is subject
to evaluation by that policy.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>