Set IMA key for container
Implement support for providing IMA keys. The keys can either be
provided via the container's Config or if none are given
the keys are (tried to be) read from the container's filesystem.
The keys are import into an _ima keyring that is created for
the container. They are loaded between the clone() and the execve()
calls and before an IMA policy is activated.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>