vtpm: Add StatepathIsManaged to VTPM JSON
Upon a 'docker restart' the state of the vTPM was deleted, which deleted
all the state of the vTPM from before. By introducing the
StatepathIsManaged variable we indicate that a higher level management
stack will delete the state path of the TPM once the container is destroyed.
This way 'runc delete' will not touch the state path anymore and leave that
up to the management stack or the user to delete the TPM's state.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>