spec: vtpm: Add parameter to start swtpm with low privileges
Extend the vtpm runtime spec with a 'runas' parameter that
allows a user to start the vTPM with low privileges.

The TPM 1.2 should be started with the user 'tss', especially when
certificates are created for it, for which tcsd is used and has to be
started either as root or as user tss. Besides that the default local
CA's lock file gives access rights only to the 'tss' user or 'root'.

The TPM 2 should also be started with the user 'tss'. Again,
also in this case the demo CA's lock file is used when certificates
are created and therefore requires to be run either as 'tss' user
or 'root'. If no certificates are created, the 'nobody' user also
works.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>